Fix a warning in the vtable for texture_2d_get_data
The function pointer for texture_2d_get_data in the driver vtable was
expecting an unsigned int for the rowstride but the definition in
cogl-texture-2d-gl.c took a size_t so it was giving an annoying
warning. This normalizes them both to just take an int. This seems to
better match the pattern used for cogl_bitmap_new_from_data and
cogl_texture_2d_new_from_data.


(cherry picked from commit 003f080531d5368835081568779b031ef4f09a77)

+++ cogl/driver/gl/cogl-texture-2d-gl-private.h
@@ -106,7 +106,7 @@ _cogl_texture_2d_gl_copy_from_bitmap (CoglTexture2D *tex_2d,
 void
 _cogl_texture_2d_gl_get_data (CoglTexture2D *tex_2d,
                               CoglPixelFormat format,
-                              size_t rowstride,
+                              int rowstride,
                               uint8_t *data);
 
 #endif /* _COGL_TEXTURE_2D_GL_PRIVATE_H_ */
